requirements.txt
geopandas
rasterio
shapely
matplotlib
numpy
pandas
scipy
torch
torchvision
torchmetrics
scikit-learn
geemap
h5py
Pillow
geopandas
rasterio
shapely
matplotlib
numpy
pandas
scipy
torch
torchvision
torchmetrics
scikit-learn
geemap
h5py
Pillow
